swarm: int is not RLP serializable, let's use uint64

+++ swarm/network/stream/delivery.go
@@ -128,7 +128,7 @@ func (s *SwarmChunkServer) GetData(ctx context.Context, key []byte) ([]byte, err
 type RetrieveRequestMsg struct {
 	Addr      storage.Address
 	SkipCheck bool
-	HopCtr    int
+	HopCtr    uint64
 }
 
 func (d *Delivery) handleRetrieveRequestMsg(ctx context.Context, sp *Peer, req *RetrieveRequestMsg) error {
